Split image settings fields
