+++ frontend/app/(core)/(workspace)/app/image/ImageWorkspace.tsx
@@ -27,19 +27,12 @@ import type { MediaLightboxEntry } from '@/components/MediaLightbox';
 import { ImageCompositePreviewDock, type ImageCompositePreviewEntry } from '@/components/groups/ImageCompositePreviewDock';
 import { buildVideoGroupFromImageRun } from '@/lib/image-groups';
 import { useI18n } from '@/lib/i18n/I18nProvider';
-import { formatAspectRatioLabel } from '@/lib/image/aspectRatios';
 import { FEATURES } from '@/content/feature-flags';
-import {
-  formatSupportedImageFormatsLabel,
-  getSupportedImageFormats,
-} from '@/lib/image/formats';
 import {
   clampRequestedImageCount,
   getAspectRatioOptions,
   getDefaultAspectRatio,
   getDefaultResolution,
-  getImageCountConstraints,
-  getImageFieldDefaultBoolean,
   getImageFieldDefaultNumber,
   getImageFieldDefaultString,
   getImageFieldValues,
@@ -59,6 +52,7 @@ import { normalizeJobSurface } from '@/lib/job-surface-normalize';
 import { groupJobsIntoSummaries } from '@/lib/job-groups';
 import { countResolvedVisualSlots } from '@/lib/group-progress';
 import { ImageLibraryModal } from './_components/ImageLibraryModal';
+import { useImageSettingsFields } from './_hooks/useImageSettingsFields';
 import { useImageWorkspaceDesktopLayout } from './_hooks/useImageWorkspaceDesktopLayout';
 import { useImagePreviewActions } from './_hooks/useImagePreviewActions';
 import { useImageReferenceSlots } from './_hooks/useImageReferenceSlots';
@@ -82,15 +76,12 @@ import { parsePersistedImageComposerState } from './_lib/image-workspace-persist
 import {
   buildCustomImageSize,
   findImageEngine,
-  formatImageSizeLabel,
-  formatQualityLabel,
 } from './_lib/image-workspace-utils';
 import {
   IMAGE_COMPOSER_STORAGE_DEBOUNCE_MS,
   IMAGE_COMPOSER_STORAGE_KEY,
   IMAGE_COMPOSER_STORAGE_VERSION,
   MAX_REFERENCE_SLOTS,
-  QUICK_IMAGE_COUNT_OPTIONS,
   type HistoryEntry,
   type ImageEngineOption,
   type PersistedImageComposerState,
@@ -168,53 +159,53 @@ export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
     selectedEngine && selectedEngine.modes.includes('t2i') && selectedEngine.modes.includes('i2i')
   );
   const selectedEngineCaps = selectedEngine?.engineCaps ?? engines[0]?.engineCaps;
-  const imageCountField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'num_images', mode),
-    [selectedEngineCaps, mode]
-  );
-  const imageCountConstraints = useMemo(
-    () => getImageCountConstraints(selectedEngineCaps ?? null, mode),
-    [selectedEngineCaps, mode]
-  );
-  const aspectRatioField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'aspect_ratio', mode),
-    [selectedEngineCaps, mode]
-  );
-  const aspectRatioOptions = useMemo(
-    () => getAspectRatioOptions(selectedEngineCaps ?? null, mode),
-    [selectedEngineCaps, mode]
-  );
-  const resolutionField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'resolution', mode),
-    [selectedEngineCaps, mode]
-  );
-  const customImageWidthField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'image_width', mode),
-    [selectedEngineCaps, mode]
-  );
-  const customImageHeightField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'image_height', mode),
-    [selectedEngineCaps, mode]
-  );
-  const resolutionOptions = useMemo(
-    () =>
-      getImageFieldValues(
-        selectedEngineCaps ?? null,
-        'resolution',
-        mode,
-        Array.isArray(selectedEngineCaps?.resolutions) ? [...selectedEngineCaps.resolutions] : []
-      ),
-    [selectedEngineCaps, mode]
-  );
-  const isResolutionLocked = Boolean(resolutionField && resolutionOptions.length === 1);
-  const supportedReferenceFormats = useMemo(
-    () => getSupportedImageFormats(selectedEngineCaps ?? null),
-    [selectedEngineCaps]
-  );
-  const supportedReferenceFormatsLabel = useMemo(
-    () => formatSupportedImageFormatsLabel(supportedReferenceFormats),
-    [supportedReferenceFormats]
-  );
+  const {
+    aspectRatioField,
+    aspectRatioSelectOptions,
+    booleanSelectOptions,
+    enableWebSearchField,
+    imageCountOptions,
+    isResolutionLocked,
+    limitGenerationsField,
+    maskUrlField,
+    outputFormatField,
+    outputFormatSelectOptions,
+    qualityField,
+    qualitySelectOptions,
+    resolutionSelectOptions,
+    seedField,
+    showAspectRatioControl,
+    showCustomImageSizeControl,
+    showEnableWebSearchControl,
+    showLimitGenerationsControl,
+    showNumImagesControl,
+    showOutputFormatControl,
+    showQualityControl,
+    showResolutionControl,
+    showSeedControl,
+    showThinkingLevelControl,
+    supportedReferenceFormats,
+    supportedReferenceFormatsLabel,
+    thinkingLevelField,
+    thinkingLevelSelectOptions,
+  } = useImageSettingsFields({
+    mode,
+    resolution,
+    resolvedCopy,
+    selectedEngineCaps,
+    setAspectRatio,
+    setCustomImageHeight,
+    setCustomImageWidth,
+    setEnableWebSearch,
+    setLimitGenerations,
+    setMaskUrl,
+    setNumImages,
+    setOutputFormat,
+    setQuality,
+    setResolution,
+    setSeed,
+    setThinkingLevel,
+  });
   const {
     canCollapseReferenceSlots,
     characterSelectionLimit,
@@ -253,180 +244,6 @@ export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
     supportedReferenceFormatsLabel,
     toolsEnabled,
   });
-  const outputFormatField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'output_format', mode),
-    [selectedEngineCaps, mode]
-  );
-  const outputFormatOptions = useMemo(
-    () => getImageFieldValues(selectedEngineCaps ?? null, 'output_format', mode),
-    [selectedEngineCaps, mode]
-  );
-  const qualityField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'quality', mode),
-    [selectedEngineCaps, mode]
-  );
-  const qualityOptions = useMemo(
-    () => getImageFieldValues(selectedEngineCaps ?? null, 'quality', mode),
-    [selectedEngineCaps, mode]
-  );
-  const maskUrlField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'mask_url', mode),
-    [selectedEngineCaps, mode]
-  );
-  const seedField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'seed', mode),
-    [selectedEngineCaps, mode]
-  );
-  const enableWebSearchField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'enable_web_search', mode),
-    [selectedEngineCaps, mode]
-  );
-  const thinkingLevelField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'thinking_level', mode),
-    [selectedEngineCaps, mode]
-  );
-  const thinkingLevelOptions = useMemo(
-    () => getImageFieldValues(selectedEngineCaps ?? null, 'thinking_level', mode),
-    [selectedEngineCaps, mode]
-  );
-  const limitGenerationsField = useMemo(
-    () => getImageInputField(selectedEngineCaps ?? null, 'limit_generations', mode),
-    [selectedEngineCaps, mode]
-  );
-  useEffect(() => {
-    setNumImages((previous) => clampRequestedImageCount(selectedEngineCaps ?? null, mode, previous));
-  }, [selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!aspectRatioField || !aspectRatioOptions.length) {
-      setAspectRatio(null);
-      return;
-    }
-    const defaultValue = getDefaultAspectRatio(selectedEngineCaps ?? null, mode);
-    setAspectRatio((previous) => {
-      if (previous && aspectRatioOptions.includes(previous)) {
-        return previous;
-      }
-      return defaultValue;
-    });
-  }, [aspectRatioField, aspectRatioOptions, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!resolutionField || !resolutionOptions.length) {
-      setResolution(null);
-      return;
-    }
-    const defaultValue = getDefaultResolution(selectedEngineCaps ?? null, mode);
-    setResolution((previous) => {
-      if (previous && resolutionOptions.includes(previous)) {
-        return previous;
-      }
-      return defaultValue;
-    });
-  }, [resolutionField, resolutionOptions, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!customImageWidthField || !customImageHeightField) {
-      setCustomImageWidth('');
-      setCustomImageHeight('');
-      return;
-    }
-    const defaultWidth =
-      getImageFieldDefaultNumber(selectedEngineCaps ?? null, 'image_width', mode) ??
-      GPT_IMAGE_2_SIZE_CONSTRAINTS.defaultWidth;
-    const defaultHeight =
-      getImageFieldDefaultNumber(selectedEngineCaps ?? null, 'image_height', mode) ??
-      GPT_IMAGE_2_SIZE_CONSTRAINTS.defaultHeight;
-    setCustomImageWidth((previous) => (previous.trim().length ? previous : String(defaultWidth)));
-    setCustomImageHeight((previous) => (previous.trim().length ? previous : String(defaultHeight)));
-  }, [customImageHeightField, customImageWidthField, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!seedField) {
-      setSeed('');
-      return;
-    }
-    const defaultValue = getImageFieldDefaultNumber(selectedEngineCaps ?? null, 'seed', mode);
-    setSeed((previous) => {
-      if (previous.trim().length) {
-        const parsed = Number(previous);
-        if (Number.isFinite(parsed)) {
-          return String(Math.round(parsed));
-        }
-      }
-      return defaultValue == null ? '' : String(defaultValue);
-    });
-  }, [seedField, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!outputFormatField || !outputFormatOptions.length) {
-      setOutputFormat(null);
-      return;
-    }
-    const defaultValue =
-      getImageFieldDefaultString(selectedEngineCaps ?? null, 'output_format', mode) ?? outputFormatOptions[0] ?? null;
-    setOutputFormat((previous) => {
-      if (previous && outputFormatOptions.includes(previous)) {
-        return previous;
-      }
-      return defaultValue;
-    });
-  }, [outputFormatField, outputFormatOptions, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!qualityField || !qualityOptions.length) {
-      setQuality(null);
-      return;
-    }
-    const defaultValue =
-      getImageFieldDefaultString(selectedEngineCaps ?? null, 'quality', mode) ?? qualityOptions[0] ?? null;
-    setQuality((previous) => {
-      if (previous && qualityOptions.includes(previous)) {
-        return previous;
-      }
-      return defaultValue;
-    });
-  }, [qualityField, qualityOptions, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!maskUrlField) {
-      setMaskUrl('');
-    }
-  }, [maskUrlField]);
-
-  useEffect(() => {
-    if (!enableWebSearchField) {
-      setEnableWebSearch(false);
-      return;
-    }
-    const defaultValue = getImageFieldDefaultBoolean(selectedEngineCaps ?? null, 'enable_web_search', mode);
-    setEnableWebSearch((previous) => previous ?? defaultValue ?? false);
-  }, [enableWebSearchField, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!thinkingLevelField || !thinkingLevelOptions.length) {
-      setThinkingLevel(null);
-      return;
-    }
-    const defaultValue =
-      getImageFieldDefaultString(selectedEngineCaps ?? null, 'thinking_level', mode) ?? thinkingLevelOptions[0] ?? null;
-    setThinkingLevel((previous) => {
-      if (previous && thinkingLevelOptions.includes(previous)) {
-        return previous;
-      }
-      return defaultValue;
-    });
-  }, [thinkingLevelField, thinkingLevelOptions, selectedEngineCaps, mode]);
-
-  useEffect(() => {
-    if (!limitGenerationsField) {
-      setLimitGenerations(false);
-      return;
-    }
-    const defaultValue = getImageFieldDefaultBoolean(selectedEngineCaps ?? null, 'limit_generations', mode);
-    setLimitGenerations((previous) => previous ?? defaultValue ?? false);
-  }, [limitGenerationsField, selectedEngineCaps, mode]);
-
   useEffect(() => {
     if (!selectedEngine) return;
     setPriceEstimateKey([
@@ -1421,84 +1238,6 @@ export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
     if (count <= 0) return null;
     return formatTemplate(resolvedCopy.messages.generatingInProgress, { count });
   }, [pendingGroups.length, resolvedCopy.messages.generatingInProgress]);
-  const imageCountOptions = useMemo(
-    () => {
-      const options = Array.from(
-        new Set<number>([
-          ...QUICK_IMAGE_COUNT_OPTIONS.filter(
-            (option) => option >= imageCountConstraints.min && option <= imageCountConstraints.max
-          ),
-          imageCountConstraints.max,
-        ])
-      );
-      return options.map((option) => ({
-        value: option,
-        label: formatTemplate(resolvedCopy.composer.numImagesCount, {
-          count: option,
-          unit: option === 1 ? resolvedCopy.composer.numImagesUnit.singular : resolvedCopy.composer.numImagesUnit.plural,
-        }),
-      }));
-    },
-    [imageCountConstraints.max, imageCountConstraints.min, resolvedCopy.composer.numImagesCount, resolvedCopy.composer.numImagesUnit]
-  );
-  const aspectRatioSelectOptions = useMemo(
-    () =>
-      aspectRatioOptions.map((option) => ({
-        value: option,
-        label: formatAspectRatioLabel(option) ?? option,
-      })),
-    [aspectRatioOptions]
-  );
-  const resolutionSelectOptions = useMemo(
-    () =>
-      resolutionOptions.map((option) => ({
-        value: option,
-        label: formatImageSizeLabel(option),
-      })),
-    [resolutionOptions]
-  );
-  const qualitySelectOptions = useMemo(
-    () =>
-      qualityOptions.map((option) => ({
-        value: option,
-        label: formatQualityLabel(option),
-      })),
-    [qualityOptions]
-  );
-  const outputFormatSelectOptions = useMemo(
-    () =>
-      outputFormatOptions.map((option) => ({
-        value: option,
-        label: option.toUpperCase(),
-      })),
-    [outputFormatOptions]
-  );
-  const thinkingLevelSelectOptions = useMemo(
-    () =>
-      thinkingLevelOptions.map((option) => ({
-        value: option,
-        label: option === 'high' ? 'High' : option === 'minimal' ? 'Minimal' : option,
-      })),
-    [thinkingLevelOptions]
-  );
-  const booleanSelectOptions = useMemo(
-    () => [
-      { value: false, label: resolvedCopy.composer.toggleDisabled },
-      { value: true, label: resolvedCopy.composer.toggleEnabled },
-    ],
-    [resolvedCopy.composer.toggleDisabled, resolvedCopy.composer.toggleEnabled]
-  );
-  const showNumImagesControl = Boolean(imageCountField);
-  const showAspectRatioControl = Boolean(aspectRatioField) && aspectRatioSelectOptions.length > 0;
-  const showResolutionControl = Boolean(resolutionField) && resolutionSelectOptions.length > 0;
-  const showQualityControl = Boolean(qualityField) && qualitySelectOptions.length > 0;
-  const showSeedControl = Boolean(seedField);
-  const showOutputFormatControl = Boolean(outputFormatField) && outputFormatSelectOptions.length > 0;
-  const showCustomImageSizeControl =
-    Boolean(customImageWidthField && customImageHeightField) && resolution === 'custom';
-  const showEnableWebSearchControl = Boolean(enableWebSearchField);
-  const showThinkingLevelControl = Boolean(thinkingLevelField) && thinkingLevelSelectOptions.length > 0;
-  const showLimitGenerationsControl = Boolean(limitGenerationsField);
   const compositePreviewEntry: ImageCompositePreviewEntry | null = previewEntry
     ? {
         id: previewEntry.id,
